fix: renamed util file isFileTree to fileTree
